refactor(@vtmn/svelte): file structure
