+++ packages/showcases/react/.storybook/preview.js
@@ -13,4 +13,9 @@ export const parameters = {
   viewport: {
     viewports,
   },
+  options: {
+    storySort: {
+      order: ['Guidelines', 'Components'],
+    },
+  },
 };
