+++ packages/sources/svelte/src/components/actions/VtmnButton/VtmnButton.svelte
@@ -1,5 +1,5 @@
 <script>
-  import { cn } from '../utils/classnames';
+  import { cn } from '../../../utils/classnames';
 
   /** @restProps { button } */
